ORM - Mapping - OneToOne

+++ 21-Spring-09-ORM-Mapping-OneToOne/src/main/java/com/cybertek/entity/Employee.java
@@ -27,6 +27,6 @@ public class Employee extends BaseEntity{
     @Enumerated(EnumType.STRING)
     private Gender gender;
 
-    private Long salary;
+    private int salary;
 
 }
